requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
configparser==7.2.0
et_xmlfile==2.0.0
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.2
propcache==0.3.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
yarl==1.20.1
asyncpg==0.30.0